Show multiple images with `.jpg` extension
Although this app could show multiple images, its speed/duration is
not stable: very fast if an image is enough small, but shows nothing
if an image is large.

If I add `std::thread::sleep` to `while` loop, window does NOT
show any images. I think that `window.update_with_buffer` is needed
to be called continuously in `while` loop. Additionally, a key input
is ignored in most cases because thread is sleeping.