Skip invalid images
If some recoverable errors are detected in the image search (init)
function, skip error image, and continue to search. The examples of
recoverable errors are:
- Invalid jpeg image
- The image which fails in saving
    - For example, too heavy to save

This function panics if no parent directory of a given photo directory
is found. In other words, when this app is run by
`cargo run --release /`, this app crashes.