Remove unnecessary attributes from error enum
`PartialEq` should not be used because there are no clear reason
to use it in the develop/release environment.

`ImageFilePathError` should not be generic because they should not
have debug message with concrete filename, path, and so on. Printing
concrete message should be managed in its caller function.

Notice that these errors aren't recoverable in this app.