Skip invalid images found during slide-show loop
If function `image_buffer_from_filepath` returns an error, its loop
is skipped using `continue` and the next image is tried to be read.

To archieve this, I introduced a new enum `ImageBufferError`. This
represents errors that could be raised in its function. Although
this error doesn't have detailed information, it's enought for this
app: the purpose of this error is to tell user of this app what they
wrong. In other words, this repository doesn't offer library.

If all images foud are invalid, this app doesn't work well. However
I think it's a very rare case.