Create `resized` dir if it doesn't exist
`resized` directory is needed to store resized images. And the
reason why resized images are needed is, to speed up image load
during slide show.

Up to a previous commit, if resize is needed and no resize directory
is here, this app fails to save resized images and could result in
an unrecoverable error (no valid images found if all images are
needed to be resized).