Prapare image buffer in another thread
With this commit, the time to read the next image is speeded up.
However, if the setting time is too short, this app might break up
because a spawned thread couldn't handle it.

Since this commit doesn't include the disconnection process: this
app panics when the app is closed. Actually this isn't a big issue
when we use this with `cargo run` command but be a serious issue
if we run tests of this app.